backend/requirements.txt
fastapi
uvicorn
pymupdf
camelot-py[cv]
python-multipart
opencv-python
pillow
pytesseract
pdf2image
pandas
pydantic
anthropic
python-dotenv
